requirements.txt
streamlit==0.87.0
numpy==1.21.2
sentence-transformers==2.1.0
huggingface-hub==0.0.12
pdfminer.six==20201018
python-docx==0.8.11
faiss-cpu==1.7.1